requirements.txt
WTForms~=2.2.1
Flask~=1.1.1
Werkzeug~=1.0.0
setuptools~=40.8.0
SQLAlchemy~=1.3.15
alembic~=1.4.2
python-dotenv~=0.12.0
email-validator~=1.0.5
flask-migrate~=2.5.3
flask-login~=0.5.0
flask-sqlalchemy~=2.4.1
flask-wtf~=0.14.3
flask-mail~=0.9.1
pyjwt~=1.7.1